requirements.txt
pytest-cov
flake8
scipy
codecov
semantic-version
